backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
sqlalchemy
pydantic
scikit-learn
joblib
pandas
groq
